added daloag while delete and ui for edit

+++ app/(dashboard)/(panel)/budgets/_components/DeleteBudget.tsx
@@ -1,10 +1,21 @@
 "use client";
 import { deleteBudget } from "@/actions/deleteBudget";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { toast } from "@/components/ui/use-toast";
 import { useAction } from "@/hooks/useAction";
 import { TrashIcon } from "lucide-react";
-import { redirect, useRouter } from "next/navigation";
+import { useRouter } from "next/navigation";
 
 type PropsType = {
   id: string;
@@ -28,20 +39,43 @@ export const DeleteBudget = ({ id }: PropsType) => {
     },
   });
   const deleteBudgetHandler = () => {
+    // console.log("hello")
     if (error) {
       toast({ title: error, variant: "destructive" });
     } else {
-      execute({ id });
+        execute({ id });
     }
   };
   return (
-    <Button
-      variant="destructive"
-      className="flex gap-2 px-2 py-3"
-      onClick={() => deleteBudgetHandler()}
-    >
-      <TrashIcon />
-      Delete
-    </Button>
+    <AlertDialog>
+      <AlertDialogTrigger>
+        <Button
+          variant="destructive"
+          className="flex gap-2 px-2 py-3"
+        //   onClick={() => deleteBudgetHandler()}
+        >
+          <TrashIcon />
+          Delete
+        </Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+          <AlertDialogDescription>
+            This action cannot be undone. This will permanently delete budget
+            and remove all related expense from our servers.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>Cancel</AlertDialogCancel>
+          <AlertDialogAction
+            onClick={() => deleteBudgetHandler()}
+            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
+          >
+            Delete
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
   );
 };
